make recovery key ui validation reusable
