+++ src/main/resources/i18n/strings.properties
@@ -445,7 +445,7 @@ recoveryKey.display.StorageHints=Keep it somewhere very secure, e.g.:\n • Stor
 ## Reset Password
 ### Enter Recovery Key
 recoveryKey.recover.title=Reset Password
-recoveryKey.recover.prompt=Enter your recovery key for "%s":
+recoveryKey.recover.prompt=Enter the recovery key for "%s":
 recoveryKey.recover.correctKey=This recovery key is correct
 recoveryKey.recover.wrongKey=This recovery key belongs to a different vault
 recoveryKey.recover.invalidKey=This recovery key is not valid
